Move values in an expanded edit row out of the way #361
Before, the first value displayed in a diagonally-expanded list of
previously collapsed values on the edit page would overlap with any
tags near the end of the row, like "!" or "changed".  Now, we have
pushed all of the values down one line so that the tags never get
overlaid and therefore never overlap visually with the value texts.