Attempt to split piles without duplicating char. data. (#254)
Comment out the code that duplicates characters and character values
because it causes the import to take longer and might be the cause
of nightly builds not completing. Try to re-use the pile suffix for
the single big pile on the two split ones. So far this seems to be
working with the exception of Get More Questions. (#254)