Fix #361 by adding length values to taxon editor
The character-value edit page for a taxon now not only shows the
multiple-choice values that are possible for the taxon, but also shows
and allows editing of length values.  This completes the refactoring
of the three character-value editing pages; they now all share common
code and logic, with the only difference being in how they are
initially rendered.