Prevent plants from being listed twice #366
A plant in two or more piles was being listed multiple times in the
plant /list/ page.

This also disables ETag for now, since to produce an ETag we either
have to run our big query twice (expensive) or else render the whole
page either way (useless).  What we need is some way for the ETag to
be computed halfway through the view's logic; I will have to look for
a Django mechanism for returning halfway through a view's logic thanks
to a matching ETag.