Remove obsolete special case from key import #382
This was an early part of the dkey logic, that is now useless because we
now have a much more thorough preprocessing step that removes empty tags
entirely; so by the time this clause runs, there are no empty tags to
special-case and skip.  I have re-run the importer with this clause gone
and done a diff on the dkey database tables without seeing any change.