Support for CSV upload of partner plant lists #376
This needs refactoring and makes me feel like we are missing a lot of
basic business-logic abstraction.  My views are full of raw database row
manipulation instead of clean assertions about what state things should
be in at the level of information.  But, the new form and logic at least
seem to work.