Add a list view option to the Level 3 page on small screens.
As a result of user tests where it was found that the usual
default thumbnails view was not as useful and desirable as
on the desktop version, especially when the results set is
large, add a list view that will soon become the default.
Add the tabs back in so users can switch to photos view
if they want. Make each entire list item an active link, and
give an appearance to invite activating the link. (#308)