Move the working area node inside the filter list item node.
This cleans up the HTML to be valid: the working area div is
inside the list item rather than after it. The list item's
click event is canceled if the event happened in the working
area, preventing the earlier problem that prompted the
decision to insert after rather than append. (#308)