Simplify Distribution and Conservation Status on species page.
Before, a wide table had states across the top and statuses across the left,
with values filled in where the existed and dashed where they did not. This
was intended to allow quick lookup and also to give a sense of what all the
available values were. However, this was wide and required horizontal
scrolling on smartphones, and user testing results suggested opting for a
simpler presentation that did not require scrolling. Now a simple list of
states with their distribution and status code is presented, which slims
down just like the Characteristics section on smartphones. (#308)