Show species-page maps higher closer to the top on small screens.
User testing found that people wanted to see the maps higher up on
the page. The maps even help a bit during plant id because people
can see whether the plant is found in their area. Before the maps
were in a 'sidebar' area pushed to the bottom as usual, but now
the maps appear up top, just below the image gallery. (#308)