Start a sample partner site 'partner' and override home page.
In the sample 'partner' home page template, so far just override the
page title, remove the Plant of the Day feed, and remove the Plant
of the Day section. (Plant of the Day can be used with partners, but
species need to be defined first.) Make a few fixes that allow the
page to display even before species are defined. (#396)