Do not add a You are Here marker on global nav. hover.
Before, if the user's cursor was left over a global navigation
item other than the current section, the page could show two
'You are Here' markers at the same time. Only one of those
markers should ever be able to show at a time so that the
user's location is always completely clear.