Avoid opening links in new browser windows.
The one place where this is tempting is the Terms link atop the
lightboxed photos, but the inconvenience of having to open one's
photo again after pressing back seems minor. If the new window
link there is really missed perhaps we can warn the user with
an icon or label.