Fix bug where Level 3 state was lost using initial list view.
Disabled code to present an initial list view on small screens,
because it was found that upon using it, state was not being
maintained when the user went to a species page and pressed
the Back button. This is not a problem when starting on the
photos view, even after switching to the list view. The bug
involves no hash being populated upon the initial list view.
(#427)