Remove unique constraint on user display name.
This was causing the import to crash upon loading fixture data,
because a blank display name was being created and then another
could not. Since it seems we may want to allow multiple
instances of the same display name (say, for users with the
same first and last names), removed the constraint for now.
However, uniqueness might be desired in the UI, which perhaps
could be achieved by pairing a display name with the user
name in parentheses, or similar. (#448)