Refine the display of PlantShare user names.
Three means of displaying user names are now available: the
regular display name, the unique display name, and the first
name. The unique display name is for situations where multiple
user names can appear on the same page, whereas the regular
display name is for use elsewhere. The first name is for
use in messages addressed to the user with an informal,
conversational tone. (#448)