Ensure Copyright (year) appears if copyright holder is blank.
The new data contain several blanks for the copyright holder field but
where other fields suggest copyright is intended. As before, the only
time the copyright notice is not shown is if the record is marked
Public Domain. This change just handles the case of the copyright
holder field being blank. (#324)