Fix #447 - Cleaner solution to dependency issue.
After some reading and experimentation on various means of dealing with
install-time dependencies, it seems that these are pretty difficult to
deal with in a way that would work for us both in our dev/setup script
AND for Heroku deployments.  Therefore, I found a fork of
django-registration which has corrected the install-time dependency and
we will point at that until appropriate patches have been pulled into
the mainline repository.