Fix issue #461.
In the event that a checklist entry had an error, but the checklist
itself did not, the code would save the checklist and then fail to save
the entry formset.  Now, we validate both the checklist form and the entry
formset before we save ANYTHING, so this should prevent duplicate
situations.