Use a link for deleting a checklist for now.
This is to be consistent with sightings, and to only allow
deleting one checklist at a time. It also allows the checkboxes
to be hidden, avoiding potential confusion given that the
feature is checklists, although we may need to revisit this
when it comes time to add the sharing feature.