First pass at fixing Literary Source admin features. (#500)
TaxonCharacterValue.literary_source replaces lit_source and is now a
ForeignKey to the new SourceCitation model. For now, the custom admin
page is still a text input, but will look up the exact text to find an
existing SourceCitation which matches.  Further improvement will change
this into a dropdown or auto-suggest box.