Fix #378 - Image moderation bugs.
The root of the issue was that on saving ScreenedImages, we were
attempting to extract GPS info each time the model was saved.  After the
intial upload however, the file was no longer local, but stored on S3, and
couldn't be opened via PIL's Image.open.  In researching the issue I
determined that GPS data extraction would be better suited to being performed
as part of the image processing pipeline provided by django-imagekit.  To get
the proper functionality required an upgrade to django-imagekit and a small
amount of refactoring of the model code.