Restrict sightings for plant names that do not match the database.
For now, unknown plants get flagged and restricted from public visibility.
Upon review by a botanist, if the plant name is corrected or a new plant
is added to the database, the user will be able to set the sighting to
public. (The original request of hiding such sightings until approval
will take some extra work.) (#526)