Remove an app name to avoid migration error.
Running a migration against plantshare caused a NoMigrations error for
me involving emailconfirmation, which is something we are using
within PlantShare. Like auth, it probably does not need to be
under South's control, and from what I can tell, we never set
it to be there. Removing it from INSTALLED_APPS should let South
skip it when running migrations, avoiding the error. (#527)