Correct some details with sighting dates.
Clean up old code to set the initial date. Change the form
field to use a datetime object instead of a date object,
even though a date is all the user can input. Set the
time behind the scenes to keep inputs ordered in the
sequence of entry on the selected date. (#535)