Fix footer issue by animating thumbnails unless on small screens.
Previously, the code would also attempt to skip animation if the
number of items was large. This caused problems with the animation
upon clearing all filters, because skipping the animation happened
on a per-filter-action basis: it did not skip based on the full
number of unfiltered results. Mixing skipping and not skipping
ended up causing trouble when clearing all. For now, animation
is not skipped unless on small screens where processor and
battery are at a premium. (#549)