Show a blank image before one is uploaded.
Change the 'No Image Uploaded' image to a blank one. The blank
space should be enough to suggest that a thumbnail will appear
there. This will also avoid any confusion in the message of the
image, as what was really meant by 'No Image Uploaded' was 'No
Image Loaded' (upload is the second step). (#566)