Allow splitting distribution records by starting letter. (#600)
Because of difficulties running the management command to split
distribution records for the entire set of plants all at once
in the Heroku environment, it was requested to add a way to
break this up by the starting letter of the plant name. This
way the command can be run a bit at a time on the plant records
so it can complete.