Upgrade to Django 1.6. (see #610)
As a last step for 1.6, South needed upgrading so that the tests
would run. Also, the test script needed changes in order to
be able to find the tests (prefixing the app name with gobotany.)
and to exclude the browser tests (suffixing the app name with
.tests). All python and browser tests are passing, although as
noted in the README, sometimes a few tests fail when run with
other apps' tests, but pass when run again alone. This latter
behavior is not new, occurring since well before the upgrade.