Fix the python-test shell script for sub-app tests. (see #612)
With the changes for Django 1.6, the script that runs the
Python tests had to be changed to surround each app name
with 'gobotany.' and '.tests' in order for the test runner
to properly find the tests. This initial change broke the
ability to run all the tests for a specific test class, or
just a single test. These are now fixed.