Upgrade to Django 1.7, including migrations. (see #613)
South is now removed, and the migrations upgraded for Django 1.7
by following the steps in the Upgrading from South section of
the Django 1.7 documentation. Instead of running the migration
commands as 'python manage.py . . .' as seen in the documentation,
we run ours using 'dev/django . . .'. And at the end of running
the migrations, got an error 'plantshare_checklist already exists',
so needed to run: dev/django migrate --fake plantshare . All seems
to be running now.