Upgrade pip and remove the distribute archive. (see #630)
Pip gave a message suggesting to upgrade to the latest. Python
appeared to be pulling down its own distribute- archive instead
of using the one committed to our repo, so it appears to no
longer be needed in the repo and is now removed. The Python
version in use during this testing was 2.7.9.