Fix bug where uploads no longer worked with current Django/imagekit. (see #645)
The error was an exception of ValueError during image upload via
the API: 'I/O operation on closed file.' The latest version of the
Django documentation on Uploaded Files and FileHandlers describes
UploadedFile.chunks() as a generator. Changing the code to use
a generator expression instead of a for loop fixed the problem.