Admin: edit textareas in monospace type, and broaden overrides. (see #592)
Before, one had to override each model class individually in order to
include the custom Go Botany admin CSS and JS files. A base class now
pulls these files in every time. So far only DKey is using it, but
the other apps can be updated to use it to allow global customizations.