Fix bug where search paging links could become too numerous. (see #647)
Before, if there was a high number of search results, the paging links
would go out to a maximum of 20, and then the list would continue to
grow in size with each additional page. Now, the paging links work
like Google's do: there are never more than the maximum number (10) of
links, and the current page appears an appropriate distance between
the first and the last.