Override D. Key admin index template to add Compact View link. (see #664)
This adds the Compact View link to the navigation of the
D. Key index view in the CMS, where the link appears in
the right side of the breadcrumbs bar. In order to do this,
the app_index.html template was copied and overridden to
add the extrahead block, which loads jQuery and the custom
JS and CSS used elsewhere.