Set STATIC_ROOT to a filesystem directory. (see #673)
The issue shows a message that talks about not setting
STATIC_ROOT to a filesystem directory, and how that is
the cause of a problem. The current configuration appears,
by setting its path to an empty string, to have /static/
appear at the root of the app directory. So, this change
attempts to keep that as it is, but express it differently
by specifying the current directory instead.