Prevent errors from appearing during Haystack rebuild_index. (see #683)
With the latest Haystack and Solr upgrade, errors started
appearing regarding too-long values in the field textSpell
for some documents, preventing them from indexing. It
turns out the site does not currently use the textSpell
field that was added some time ago in order to help with
spelling suggestions. After removing the field, running
a query on a misspelled word in the Solr Admin still
came up with spelling suggestions (spellcheck), so it
seems using those is still an option for the future.