Fix bug where upload dialog did not show thumbnail for small images. (see #694)
This bug would not be seen often on the site, and does
not affect the issue with image upload errors. It
happened only for images that were smaller than our
maximum size, which was seen in a small test image.
Most images, such as from a digital camera, would not
have triggered the bug.