Update to the latest Django 1.11 release, .18, fixing an Admin bug. (see #714)
The poor performance that was the cause of this issue was
largely due to the Django bug introduced in one of the
recent releases (probably a 1.11.x release), which led to
Django bug #28462 being logged. The fix was included with
the 1.11.14 release. Now saving checkbox edits on the
Distribution list view takes a respectable ~11 seconds
again, whereas before the fix it was taking ~50 seconds,
too long for Heroku which times out at 30 seconds.