Fix search bugs by switching to an earlier version of Haystack. (see #717)
2.7.dev0 is the earliest that does not complain that it is
incompatible with Django 1.11. 2.7.dev0 does not exhibit the
bugs seen in #717, which consists of strange and broken
behavior with results queries for some queries. Versions
later than 2.7.dev0 (2.7.0 and on) do exhibit the bugs.
Versions back to at least 2.5.1 do not exhibit the bugs,
but setup complains about incompatibility with Django 1.11
(it does seem to work fine there, however).