Make the D. Key Admin's Lead change page faster. (see #744)
A major problem was that the editable foreign key field
Goto_Page has too many values to show a regular select
field for it, and the querying alone for it caused the
page to fail to appear. Using raw_id_fields relegates the
value selection to a popup dialog, making the page faster.